CharacterController component - saves data to scene

+++ Engine/Graphics/Vulkan/Rendering/VulkanDeferredLight.cpp
@@ -156,7 +156,7 @@ void VulkanDeferredLight::RecordCmdbuf(VulkanCommandBuffer* cmdbuf) {
 
 	_deferred_rp->CmdBegin(*cmdbuf, *_deferred_fb);
 	cmdbuf->BindPipeline(*_deferred_pipeline);
-	cmdbuf->SetViewport(0, 0, _fb_width, _fb_height);
+	cmdbuf->SetViewport(0, 0, (float)_fb_width, (float)_fb_height);
 	uint32 cam_offset = _camera_index * CAMERA_ELEM_SIZE;
 	cmdbuf->BindDescriptorSets(*_deferred_pipeline_layout, 0, 1, _deferred_descriptor, 1, &cam_offset);
 	cmdbuf->BindMesh(*mesh, 0);
